RB
Har fixat settingsmeny i gamemode, audio bakgrund, så bakgrundsbild hamnar rätt
