--- BActeria Destruction 3D/Assets/HazardBomb.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 55717489b9d8cae4aae0bed9731a390e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
