+++ BActeria Destruction 3D/ProjectSettings/TagManager.asset
@@ -39,6 +39,12 @@ TagManager:
   - 
   - 
   m_SortingLayers:
+  - name: Background
+    uniqueID: 3842814845
+    locked: 0
+  - name: Menu
+    uniqueID: 2317043007
+    locked: 0
   - name: Default
     uniqueID: 0
     locked: 0
